add sweep function and tests, rename to EmissionsDripper, add comments
