+++ tests/deployment/test_emissions_dripper.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 import pytest
-from brownie import Contract, accounts, chain
+from brownie import accounts, chain
 from brownie_tokens import MintableForkToken
 
 from helpers.addresses import registry
@@ -14,7 +14,7 @@ def deployer():
 
 @pytest.fixture(scope='module')
 def dripper():
-    from scripts.deployment.deploy_rembadger_dripper import main
+    from scripts.deployment.deploy_emissions_dripper import main
     return main()
 
 
@@ -25,6 +25,7 @@ def badger():
 
 @pytest.fixture(scope='module', autouse=True)
 def topup_and_fast_forward(dripper, badger):
+    accounts[1].transfer(dripper, 1e18)
     badger._mint_for_testing(dripper, 100_000 * 10**badger.decimals())
     assert badger.balanceOf(dripper) > 0
     now = datetime.now().timestamp()
@@ -92,3 +93,27 @@ def test_set_keeper_from_techops(dripper, techops):
 @pytest.mark.xfail
 def test_set_keeper_from_random(dripper):
     dripper.setKeeper(accounts[1], {'from': accounts[1]})
+
+
+def test_sweep_eth_from_dev(dripper, dev):
+    bal_before = dev.account.balance()
+    dripper.sweep({'from': dev.account})
+    assert dev.account.balance() > bal_before
+    assert dripper.balance() == 0
+
+
+@pytest.mark.xfail
+def test_sweep_eth_from_random(dripper):
+    dripper.sweep({'from': accounts[1]})
+
+
+def test_sweep_erc_from_dev(dripper, badger, dev):
+    bal_before = badger.balanceOf(dev)
+    dripper.sweep(badger, {'from': dev.account})
+    assert badger.balanceOf(dev) > bal_before
+    assert badger.balanceOf(dripper) == 0
+
+
+@pytest.mark.xfail
+def test_sweep_erc_from_random(dripper, badger):
+    dripper.sweep(badger, {'from': accounts[1]})
